requirements.txt
streamlit==1.34.0
pandas
SQLAlchemy==1.4.0
pyodbc
matplotlib
streamlit-modal
altair
streamlit-extras
pygithub